Add explicit=True to TransactionManager
This tells our transaction manager to crash, rather than implicitly
creating a new transaction, if something tries to access the DB after
the transaction has ended.

AFAIK we don't have anything that tries to access the DB after the
transaction has ended but, given that doing so could cause unclosed DB
connections to be silently leaked, this is a precaution to prevent
future changes from doing that.

pyramid_tm recommends doing this:

https://docs.pylonsproject.org/projects/pyramid_tm/en/latest/#custom-transaction-managers

For how unclosed DB connections could be silently leaked without this,
see:

https://github.com/hypothesis/h/pull/5702#issuecomment-531243297